Move expectJobWithSuspendedAndNodeSelectors to test and remove util.go
